+++ R/utility.R
@@ -78,4 +78,5 @@ get_timestamp <- function() {
   make_svm_rbf_multistep()
   make_xgboost_multistep()
   make_timegpt_model()
+  make_chronos2_model()
 }